[FL-1513] Fix IRDA crush
There was an additional assert added into ISR callback setup, checking
that callback was NULL if you set it, and not-NULL if you reset it.
IRDA always calls for api hal irda deinit function in destructor in
spite of was it already done or not. So, add flag to check this
condition and don't perform double periphery deinitialization.