Change from one big command to three small commands
These changes follow from the review. Instead of having to
rely in GOTOs (which are forbidden :<( ) to ensure cleaning,
I divided in smaller pieces. Since I had no feedback w.r.t
the previously existing `gpio_set`, this fits more with the
current implementation.

I also tried to interpret "That's not how we do enums" as
to not include a `enum EnumName` decl. when using typedefs.

I also distanced the implementation from using LL_GPIO
in favour of functions provided in `<furi_hal_gpio.h>`
but there was nothing to get the mode, so I stuck with
LL_GPIO for mode check.