Added UPC-A Barcode generator.
This generator has no safety, it doesn't preform any checks on the internal logic of the barcode, it simply converts the 12 digit number at the bottom in to the UPC-A format of barcode. Hitting the centre button toggles between "Edit" mode. Left and right change the cursor position, and up and down to change the current digit.