Add Nix expressions to build firmware with Nix
[Nix](https://nixos.org/) is distro-agnostic package manager that can also be seen
as a meta build system.
This can be used to build on any Linux distro as long as the Nix
package manager is installed, and possibly also on MacOS though I
haven't tested this myself.

This is intended to be used either purely for building the firmware.
Building like this runs the build inside a sandbox.
- `$ nix-build ./.`

Or interactively using nix-shell:
- `$ nix-shell`
- `$ ./fbt firmware_flash`