Allow extended frequency range
Added a secret settings file in SubGhz/Assets that enables the same frequency range as a Yard Stick One. This is use at your own risk, so have a look in the firmware to find the name of the file and the key you need. The file is the same format as settings_user