Increment U2F counter before using it
The spec at [1] is not entirely clear on what the initial counter value
should be:

> A U2F token must increase a counter each time it performs an
> authentication operation.
> [...]
> U2F token counters should start at 0.

[1] https://fidoalliance.org/specs/u2f-specs-1.0-bt-nfc-id-amendment/fido-u2f-implementation-considerations.html#token-counters

And [2] does not mention the initial value at all.

[2] https://fidoalliance.org/specs/fido-u2f-v1.2-ps-20170411/fido-u2f-overview-v1.2-ps-20170411.html#counters-as-a-signal-for-detecting-cloned-u2f-devices

However, the auth flow at [3] shows that the counter is incremented
_before_ signing/sending:

u2f-device     client   relying-party
      |           |           |
      |           |<-challenge|
      |<-challenge|           |
cnt++ |           |           |
      |cnt,sig()->|           |
      |           |cnt,c,s--->|

[3] https://developers.yubico.com/U2F/Protocol_details/Overview.html

Behaving the same as the Yubico flow is compatible and safe because
incrementing once too many is never a problem.

This fixes triggering a Cloned-Device detection as seen in [4] where the
relying-party counter has 0 stored and expects 1 as the legit counter
value.

[4] https://github.com/ossobv/django-otp-u2f/blob/33f38dbc5dfb01a4c56e82707fe7168f75f1b3fa/otp_u2f/models.py#L111-L127