lfrfid: decode HID size header and render it
Format: the HID Proximity data is 44 bits after Manchester-decoding. The first six bits seem to be zero but there is not much documentation on this. The seventh bit starts to indicate the length. If the seventh bit is `0`, the remaining 37 bits make up the key data: this is what people call "37-bit HID". If the seventh bit is `1`, all of the bits until (and including) the next `1` are part of the header and the remaining bits make up the key data.

This commit decodes the header to infer the key size. If the key size is under 26 bits, the minimum HID key size AFAIK, the size is set to `HID_PROTOCOL_SIZE_UNKNOWN`.

This commit also displays the key data without the header. It does not parse the data further to display fields like facility code, card number, and parity bits.

The code is written for readability more than speed, especially when rendering the key data, but iterating over 37 bits to display them should not be a bottleneck.

(I also changed `size_t protocol_size` to `uint8_t protocol_size` because we are guaranteed the HID protocol size is at most 37 bits.)