lfrfid: decode generic HID Proximity size lazily (#1618)
This fixes a bug when displaying info about a captured HID key on the Flipper. Previously, `protocol->protocol_size` was set when decoding raw bits. However, the size was not set when loading a key that was already decoded and saved to disk (or a manually entered key).

This commit lazily computes the key size with a new function named `protocol_hid_generic_decode_protocol_size` and removes the `protocol_size` from the HID protocol struct.

Tested this fix by loading a saved HID key on the Flipper and verifying that the Flipper shows the key size.