Fix intermittent crash caused by race condition
This fixes a race condition in which the IRQ cb would reference thread data before it was ready, causing a NULL pointer dereference.
This happened when receiving UART data early in app startup, which is why it could be reproduced only with a board attached.
More investigation is needed to determine why/if the USB-UART bridge in the main GPIO app isn't affected (which has a similar architecture).