NFC: Fix MIFARE DESfire info action to open app menu
When examining a MIFARE DESfire card, selecting "More >" currently leads
to the file menu for a single application even if the scanned card
contains multiple applications.

On examining the source code, a MIFARE DESfire application selection
menu is present as `NfcSceneMfDesfireData`. This change updates the
MIFARE DESfire Info "More >" action to open the application selection
menu. That menu may then be used to open the file selection menu for any
application in the read MIFARE DESfire card data.

Tested interactively with
https://github.com/flipperdevices/flipperzero-sd-card-examples/blob/c4cbdcd94706836fe0e3cda42587814237b1cfde/nfc/Desfire.nfc