power: Add unit tests for battery charging voltage
Add unit tests to verify that the battery charging voltages are
properly set and avoid unsafe values.  This might require the Flipper
to be plugged into USB power (e.g. the USB connection for testing).

NOTE: These tests involve modifying the charging voltage for the
battery.  If they fail, it's a good idea to make sure the charging
voltage isn't stuck at a dangerously high value.