Power: fix battery charge voltage limit for > 7.935v
Fix uint8_t wrap-around for battery charge voltage limits > 7.935v.
The type casting would incorrectly result in a value such as 8.105v
resulting in a value below 23 (e.g. 10), which is valid but probably
not the intent.

Setting 4208 as the upper bound ensures that passing in 255.0f to
FuriHal picks the highest valid charge voltage limit, not some number
between the highest and lowest bounds.

The after-casting VREG comparison is kept for safety reasons.  It
should be impossible to hit this, but when dealing with lithium ion
batteries it's probably better to be safe than sorry.