ibutton: Fix DS2502 reading and emulation
DS2502 adds a leading checksum byte using Maxim CRC8 before pushing out any actual bytes read. This breaks reading and emulation as the checksum was not handled in the current read and emulate routines. An option has been implemented for both dallas_common_read_mem and dallas_common_emulate_read_mem to be aware of this checksum if the device needs it to be read and emulated properly.