Bring back Lite data viewer and add monolithic NDEF support
Monolithic NDEF tags are similar to Lites as they don't respond to most commands. The IC code is 0xff and default system code is 12fc. They usually respond to read requests on read-only open random service #0 and seems to terminate the read with an access denied status code instead of invalid block address.

The Lite data viewer is largely based on @nullableVoidPtr's work.