New card detection routine
This allows the reader to try out common monolithic systems (i.e. tag doesn't respond to Request System Code but has 1 common systems on the tag) (currently handles Lite and NDEF) when the tag doesn't support Request System Code and the IC code is unknown to the reader.

Things that are left out from this commit and are in future TODO:
- Improve performance by checking obvious IC codes in the magic detection routine
- Update the FelicaData struct to include a monolithic system flag
- Update felica_print_card_stat to depend on monolithic system flag and system code instead of IC code to determine what stat format to use
- Make the NFC stack respect to PMm (low priority, need rebase)