pokemon: Refactor party trade struct setup/use
No longer declared in header, causes problems for multiple
inclusions. Added as a pokemon_fap struct member.

Trade was updated to use this struct member. For sections of code
where the flat array is still expected, the TradeBlock struct
is cast as a flat pointer.

The struct initialization is also cut down heavily to only specify
the bare minimum needed for the first pokemon in the party.