Updated index.html for Modern Captive Portal Login (#18)
* Updated index.html for Modern Captive Portal Login

Everything is embedded in the HTML file. The CSS is included within the <style> tags in the <head> section of the HTML, and there is no JavaScript in the code provided. This HTML file is completely standalone and does not require any external CSS or JavaScript files to function.  because of limitation that all of the HTML/CSS/JS must be in a single HTML file that does not exceed 20k character

* Created community portals folder, migrated portal

---------

Co-authored-by: bigbrodude6119 <138256922+bigbrodude6119@users.noreply.github.com>