pokemon: Refactor, UI update, feature update
This patch moves the UI system to using a scene manager as well
as a few builtin modules as opposed to a new view for every
step. This decreases the binary size as well as presumably
reduces runtime memory footprint.

It also folds back in the features added in #17 and manages
to fix an experience calculation bug. Note that, the features
from #17 were not fully tested and there may be some incorrect
differences at this time.

This commit adds new features on top of that, the ability to
specify a custom nickname for the traded pokemon, its type(s)
as well as its OT name and ID#.

Types were modified a bit from the original implememtation in #17.
In Gen I, a pokemon with a single type just repeats that type
twice in its data structure. Rather than using 0xFF as a "No Type"
indicator, the main data table repeats types for single typed
pokemon.

There is a bit of a shortcoming with setting types. There is no
way to easily reset to default types. Due to the way the item_list
works, it directly modifies the actual trade data block. The only
way to reset at the moment is to select a different pokemon, and
then re-select the desired pokemon.

The solution to the Nidoran issue was reworked a bit. I had
issues compiling with the unicode escape characters, while
others had issues compiling with the actual unicode characters.
It would have been possible to use wchar_t for the pokemon_table
names, however that would have quadruled the size of each string.
Regardless, the Flipper can't print the male/female symbols at
this time. It might be possible to hack them back in with some
custom icon and View manipulation. The fix was to just use extended
ASCII character values. Since the Flipper doesn't support those
either, it still just renders as a "space". It keeps the strings
as char arrays. Compiles in any environment. And works.

Some of this implementation is still quite hacky and could use
an overhaul before new feature additions. e.g. Adding Gen II
support might not easily work with the current state of the
codebase.

Fixes #7
Fixes #8