Latest Release RM1012-1745-0.93.1-ae9f9be on PATREON - THANK YOU TO ALL SUPPORTERS AND COMMUNITY CONTRIBUTORS!
Well it looks like I won't be in the unleashed discord anymore. MMX/xMasterX/Nano decided to ban me after contributing $1 to the project. In the past, I have given several sizeable BTC donations to the unleashed project and it was received with disrespect. Therefore, I gave a $1 donation to see how it would be treated since it was an option, I guess he didn't like it. Anyone can see what transpired in his general chat unless he wiped it clean out of shame. Most of his discord channels' chats' were other devs trying to poach users or beginners getting shamed for not knowing any better, it was kind of frustrating to watch. Although unleashed is a great project and the work of xMasterX is wonderful, he is a shit person.
I do what I do for this firmware for the enjoyment of the device, I don't have any expectations of support from our users and give everyone the opportunity to have the best firmware available. Pre-compiled for supporters daily, available to be compiled by all at any time and pre-compiled for all users weekly. RM firmware will not change and we will continue to get all new updates and features from the open source community including all the features from the unleashed repository. Contrary to the slander out there, credit will be given for all contributions as they always have been. However, no longer will there be an open advertisement for unleashed on our ReadMe as was requested by xMasterX. It is unfortunate that one of our community devs can be so petty.