trade: added some robustness when going back to menu
Still has some crash for some reason when trying to re-enter trade
some times. Need to actually debug that.

Also cleans up the trade re-entry by forcing the state back to
reset_init and respond to trade requests in the ready state with
a cancel. Allows the gameboy to also back out and then re-enter
the trade screen and re-sync everything.