Fix: add_view must be called on all scenes
The tutorial I based my code on was wrong. The types of
modules used in the program can be less than the number
of views (or scenes), BUT with view_dispatcher_add_view
we are telling that one view (or scene) is going to use
that module type.

It's wrong to consider a different enum for the modules
types and use that to register the views.