Don't create blank list entry for newline at EOF
If a question list is edited using Linux, a newline is appended to the end of the file.  The generate_questions.py script interprets this as a list entry, as we end up with a blank list entry on the flipper.  This fix checks for a newline at the end of the file and removes it if necessary.