Fix: Update check for dynamic release tag in workflow
The build workflow for 'Subbrute' application was updated to use a dynamic check for the release tag in Github API call. Additionally, we are now fetching the SHA, firmware version and URL, which makes debugging easier. This enhancement increases efficiency, reduces unnecessary API calls and mitigates potential issues if a specific release tag doesn't exist.