Setup Main Menu
[Problem]
Currently the application does nothing.

[Solution]
This updates the app to have a simple main menu, with the options
that will be available in the final version of the app. It also
sets up some directory structures, dividing the code up into
pieces that are more easily digestible. Specifically, different
scenes can be managed on their own, without being as dependent on
a single file.

[Testing]
Built and confirmed working on device. Should probably investigate
unit testing but that can come later.