Re-architect View Management + Add Settings Scene
[Problem]
Attempting to add a settings scene, but it turns out that
removing an in-use view by the view dispatcher causes some pretty
gnarly issues.

[Solution]
As a result, it is better to instead share views
between scenes for simplicity. This is a small sacrifice in heap
memory for simplicity, though maybe this can be re-addressed later.

[Testing]
Confirmed working on device.