Add Playback Scene + Tone Properties
[Problem]
The playback button doesn't do anything at the moment.

[Solution]
Added the playback scene which displays a waveform based on
the properties of the tone. Properties of the tone are stored
in the application context and then passed to the view model
that displays the animation.

The view for the animation is simply a custom view that has
a callback called to draw to its canvas continuously. The waveform
is drawn with lines in order to ensure it is drawn properly connected.
Can currently display sine waves and square waves.

[Testing]
Ran on device and confirmed working as expected.