Update App Context to Add Views to Dispatcher + Add SUpport for Variable Item Lists
[Problem]
Right now, I need to add views to the dispatcher in the main code, even though
all of the properties are in the view when added to the app context.

[Solution]
Updated the app context to add views to the view dispatcher, and also added
support for variable item lists.

[Testing]
Tested on device and confirmed everything still runs.