Switch to Frequency + Rework Settings for Variable Item List Use
[Problem]
Currently using period and frequency to generate the tone to play.
This isn't easy to understand, and isn't nicely compatible with
understanding what tone will play.

[Solution]
Updated to use frequency for the tone to play, plus updated the
Settings Scene to instead use a variable item list for changing
the settings. This makes the settings manageable now.

[Testing]
Tested on device and confirmed working.