Started implmentation of custom StartScreen view/module that can be reused for other games.
Right now the view model has 3 text elements and an icon, maybe expand to add arbitrary UI elements like widget does in future?

I realized it is similar in function to the Popup module right now but has the intention of allowing more modification of the view callbacks and UI elements. Adding UI components is similar to widget, and this implementation may be kind of a redundant view/module in general but gives me the flexibility to customize the starting screen view of the app however I want.