Started skeleton for MineSweeperGameScreen view
This view/module with have the actual game state.

I think removing the public methods for the input callback is a good idea as it should all be defined
within the game view and should rely one management within the scene.

This could allow the module to be exportable to other apps.

This means that all checking and game logic needs to happen internally with minimal access outside of the view.
Ideally the scene will just handle the custom event types passed down by this view when appropriate.