Redefined temporary macros for width and height to test
We allow 1024 tiles max right now; the ideal implementation will just have the macro for MAX_TILES and SCREEN_TILE_WITDH/HEIGHT and the board
width and height will be controlled by the user in an options menu accessible to the user from the back screen.

STARTING_MINES macro could be replaced in this menu with difficulty levels where each level is a different multiplier to the total tile count used.