Added new functionality for variable size board with ability to scroll through the grid.
A uint8_t "boundary" variable was added for the bottom and right side of the current slice
that we can view of the total game grid. A more detailed explanaion may
be added later.

Game draw function updated to work with this boundary to transform the
relative slice of the game view to the absolute coordinates on the game
grid.

Other variables were upgraded from uint8_t to uint16_t to allow for a
larger board technically up to 2^16 tiles, but we limit to some value
like 1024 with a macro in the .h file.