Added initial skeleton for settings scene
Right now it has one list item for difficulty and a callback to cycle through options.

This screen should read and populate its data from the settings_info member in the main app struct on enter, and on exit should commit those changes to the struct.

This means we will probably need another menu screen for confirmation if they want to save the data..