Added callbacks for item change in each menu item & added event enum
Each event callback handles the change and modifies the settings_info struct within the main app struct. For the width and height item change
callbacks I needed a reference to the other item to be able to modify it so when increasing the w/h the value of w*h was not greater than
the max allowed tiles.