Fixed back button bug for start screen
Not sure if something changed with the new firmware, but pressing
or holding the back button on the start screen would lead to null-ptr
dereference when it didn't before. Removing the logic that actually
handles the back button press and only processing non back button inputs
probably allows for the default back-press callback to be used which exits the
app from this view as intended.