views/trade: Some hacks in place to connect to gen ii
Can't trade yet, and its definitely quirky, but am able to swap
party info and verify that data.

trade: fix incorrect patch list start offset

Interestingly, this did not seem to affect the actual trade, without
looking at pokemon disassembly, I guess its not counting the number
of 0x00s and rather looking for something nonzero or 0xFF to end
the part. 0x00 is a blank byte, 0x01 means to patch the 0th byte.

trade works from flipper to GB, GB to flipper broken

Flipper doesn't correctly update itself, it also does not go to
trade animation, instead back to ready. Repeated table use causes
the flipper to quickly go waiting and then ready again

trade: More updates to support gen ii trades

Currently trade succeeds. There are some broken issues with leaving
and joining the table in a safe way still.

Reject a trade is functional though

trade: Most everything is functional, needs clean up

Need to work on the main state machine, its possible still to desync
the flipper and gb if both enter the table/waiting, the flipper backs
out, and re-enters trade, and the gameboy attempts to start a trade.
The flipper will end up responding by trading the same numbered pokemon
that the flipper did.

The issue is that gen ii introduces some additional header that looks
the same as a trade request. So, need to synchronize states a bit,
somehow.

trade: hacks to better support gen ii

This is a hacky fixup of the synchronization issues that occur
when the flipper backs out of the trade but the game boy does
not.

The result now is the gameboy just kind of resetting the trade
state to the main trade menu. Canceling and re-entering the trade
table resynchronizes all of this.

trade: account for reused meny selection byte