Fixed error in bfs_tile_clear_verifier(); other minor edits
There was a slight error where we were not continuing when the current position
was in either of the visited sets. Hopefully this fix will decrease the time
it takes for the verifier to check a board as it is no longer redundantly checking
tiles we already visited in the bfs tile clear.