trade: Most everything is functional, needs clean up
Need to work on the main state machine, its possible still to desync
the flipper and gb if both enter the table/waiting, the flipper backs
out, and re-enters trade, and the gameboy attempts to start a trade.
The flipper will end up responding by trading the same numbered pokemon
that the flipper did.

The issue is that gen ii introduces some additional header that looks
the same as a trade request. So, need to synchronize states a bit,
somehow.