trade: hacks to better support gen ii
This is a hacky fixup of the synchronization issues that occur
when the flipper backs out of the trade but the game boy does
not.

The result now is the gameboy just kind of resetting the trade
state to the main trade menu. Canceling and re-entering the trade
table resynchronizes all of this.