Implemented wrapping player movement in game view
Added ability for the cursor position to wrap to the other side of
the game grid instead of blocking at the edge. Data member was added tothe model,
and is set within alloc and the set_wrap_enable function (added
as a public api to allow other scenes to set the wrap enable when it
is changed in places like the settings screen).

I also refactored the player_move function so a lot of the logic isn't
repeated in each switch statement and the border boundaries for the
displaying of the board is generalized when wrapping.