countdown timer: make digit selection wrap when pressing right
When pressing right, adds the modulo number to the `selection` variable
before subtracting and modulo'ing it. This prevents it from being
assigned a negative value (which cannot be relied upon for an enum),
as well as making the variable properly wrap around below "zero".

While not strictly necessary, the code for handling left button presses
is also changed for consistency.