lib: digital_signal: digital_sequence: add furi_hal.h wrapped in ifdefs
Per the comment at the top of the file, defining DIGITAL_SIGNAL_DEBUG_OUTPUT_PIN
to be a GpioPin variable name should allow additional debug output on
that pin. However, this would not work without modifying the file as
well to add the furi_hal.h header. Wrap including that header in the
same macro define to automatically include it when used.

Fixes: d92b0a82cc4c ("NFC refactoring (#3050)")

Signed-off-by: Kris Bahnsen <Kris@KBEmbedded.com>