gpio: clear irq status before calling user handler
Previous behavior was the following for EXTI IRQ handlers:

    if (irq_enabled) {
        call_user_handler();
        clear_irq_status();
    }

This is a problem if the EXTI line changes during the IRQ: presently,
the IRQ status is effectively logically unobserved during the user
handler, as the status is unconditionally cleared afterwards.

You can see this by building a simple ring oscillator circuit: feed an
output pin back into an input pin and invert the input level in an
interrupt. The circuit doesn't oscillate without this patch, as the
IRQ triggers before the status is cleared. With this patch, the circuit
does oscillate.

This change inverts the order of the logic to clear the IRQ status
first.